INFUND-6923 - Solved type incompatibilities after merge, and populator can now handle restfailure gracefully
Former-commit-id: c4257b7cb4288e6fceb43c54105d483b368f77bd

+++ ifs-web-service/ifs-application-service/src/main/java/org/innovateuk/ifs/competition/CompetitionSearchController.java
@@ -28,7 +28,7 @@ public class CompetitionSearchController {
     @GetMapping
     public String publicContentSearch(Model model, @RequestParam(KEYWORDS_KEY) Optional<String> keywords,
                                       @RequestParam(INNOVATION_AREA_ID_KEY) Optional<Long> innovationAreaId,
-                                      @RequestParam(PAGE_NUMBER_KEY) Optional<Long> pageNumber) {
+                                      @RequestParam(PAGE_NUMBER_KEY) Optional<Integer> pageNumber) {
 
         model.addAttribute("model", itemSearchPopulator.createItemSearchViewModel(innovationAreaId, keywords,  pageNumber));
